Let installer fixtures express executable modes (#716)
The POSIX resolver regression was locked with a hand-built pair while the shared installer wrapper always chmodded every fixture 0755. That left the main fixture API unable to construct the state the regression depends on, so a future test using the helper could still produce a misleading green.

Move the shadow assertion beside the shared wrapper and let the wrapper take an explicit mode. Its default remains 0755, while the regression now asks the ordinary fixture path for 0644 and 0755 files.

Limit: test fixtures only; production resolution and spawning are unchanged
Ruled-out: keep the hand-built shadow pair | it preserves the helper blind spot the owner identified
Blast: local
Undo: easy
Certainty: firm
Record-Id: r-fixture716
Provenance: authored
Verified: tsc --noEmit clean; Windows installer suites 29 passed with the one POSIX-only executable-mode case skipped
CommitLore-Version: 2.0.0
